Added team validation
Added protection against bad requests & lists all matches by team.
¡¡WARNING!! This is not final, because still waiting TeamService bearer resolution.

+++ match_service/match_api/views.py
@@ -13,22 +13,31 @@ OPEN_WEATHER_KEY = "OPEN_WEATHER_KEY"
 DEFAULT_OPEN_WEATHER_KEY = 'b8de83b3476d58590a4fbf3661f4dabe'
 TEAM_SERV_URL = "TEAM_SERV_URL"
 
+team_backend_url = os.getenv(TEAM_SERV_URL, 'https://team-service-danaremar.cloud.okteto.net/api/v1/')
+
+def validate_token(headers):
+    return requests.post(team_backend_url + ' validate-token', headers=headers)
+
 class MatchViewSet(viewsets.ModelViewSet):
 
     serializer_class = MatchSerializer
 
     # list
     def list(self, request):
-        # TODO: access to TeamService
 
-        queryset = Match.objects.all()
+        # check permissions
+        bt = validate_token(self.headers)
+        if(bt['cod']!=200):
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        
+        # list all matches from team
+        # queryset = Match.objects.all()
+        queryset = Match.objects.filter(user_uid=bt['pk'])
         serializer_class = MatchSerializer(queryset, many=True)
         return Response(serializer_class.data)
         
     # get
     def get(self, request, pk=None):
-        # TODO: access to TeamService
-
         match = self.get_object(pk)
         serializer_output = MatchSerializer(match)
         return Response(serializer_output.data)
@@ -36,38 +45,42 @@ class MatchViewSet(viewsets.ModelViewSet):
     # create
     def create(self, request):
 
-        # TODO: access to TeamService
+        # check permissions
+        bt = validate_token(self.headers)
+        if(bt['cod']!=200):
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         open_weather_key = os.environ.get(OPEN_WEATHER_KEY, DEFAULT_OPEN_WEATHER_KEY)
 
         serializer = MatchSerializer(data=request.data)
 
         city = request.data['city']
         api = requests.get('https://api.openweathermap.org/data/2.5/weather?q=' + city + '&appid=' + open_weather_key).json()
-        #celsius = (api['main']['temp'] - 32) / 1.8
         
         if api['cod'] == '404':
             return Response(status=status.HTTP_400_BAD_REQUEST)
 
-
         if serializer.is_valid():
-
             celsius = api['main']['temp'] - 273.15
-            
             serializer.validated_data['weather'] = api['weather'][0]['description'] + ' - Temperatura: %.2f ' %celsius
             serializer.save()
-            
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
             return Response(serializer.data, status=status.HTTP_400_BAD_REQUEST)
 
 
-        
-
     # update
     def update(self, request, pk):
-        # TODO: access to TeamService
 
+        # get saved match
         match = self.get_object(pk)
+
+        # check permissions
+        bt = validate_token(self.headers)
+        if(bt['cod']!=200 and match.user_uid == bt['pk']):
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
+        # update if valid
         serializer = MatchSerializer(data=request.data)
         serializer.is_valid()
         serializer.update(match, request.data)
@@ -75,9 +88,15 @@ class MatchViewSet(viewsets.ModelViewSet):
 
     # delete
     def delete(self, request, pk):
-        # TODO: access to TeamService
-
+        # get saved match
         match = self.get_object(pk)
+
+        # check permissions
+        bt = validate_token(self.headers)
+        if(bt['cod']!=200 and match.user_uid == bt['pk']):
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        
+        # delete
         match.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -96,8 +115,7 @@ class SendEmailSet(viewsets.ModelViewSet):
 
         for match in matches:
             # TODO: send to TeamService
-            team_service_url = os.getenv(TEAM_SERV_URL, 'https://team-service-danaremar.cloud.okteto.net/api/v1/')
-            requests.post(team_service_url + 'send-email-player', data=match)
+            requests.post(team_backend_url + 'send-email-player', data=match)
 
             # updated with sent_email
             modified_match = match
